requirements.txt
Flask==2.3.3
argon2-cffi==23.1.0
cryptography==42.0.5
qrcode[pil]==7.4.2
requests
python-dotenv
zxcvbn
pytest
gunicorn
